Configured README file and include an example of a compose one

+++ src/main/kotlin/com/sinaglife/shoping_cart/cart/domain/DiscountApiClient.kt
@@ -5,7 +5,7 @@ import com.sinaglife.shoping_cart.cart.domain.cart_discount.CartDiscount
 interface DiscountApiClient {
     fun getDiscount(
         code: String,
-        subTotal: Int,
+        subTotal: Double,
         customerId: String?
     ): CartDiscount
 }
