+++ src/main/kotlin/com/sinaglife/shoping_cart/cart/domain/cart_item/CartItemPrice.kt
@@ -1,4 +1,4 @@
 package com.sinaglife.shoping_cart.cart.domain.cart_item
 
-class CartItemPrice(val value: Int) {
+class CartItemPrice(val value: Double) {
 }
