+++ src/main/kotlin/com/sinaglife/shoping_cart/cart/domain/events/CartCreatedDomainEvent.kt
@@ -8,9 +8,9 @@ import java.time.LocalDateTime
 data class CartCreatedDomainEvent(
     val id: String,
     val items: List<CartItemPrimitives>,
-    val subTotal: Int,
-    val total: Int,
-    val discount: Int?,
+    val subTotal: Double,
+    val total: Double,
+    val discount: Double?,
     val createdAt: LocalDateTime,
     val updatedAt: LocalDateTime,
     val customerId: String?,
