+++ src/main/kotlin/com/sinaglife/shoping_cart/read_model/application/search/SearchCartQuery.kt
@@ -8,9 +8,9 @@ import java.time.LocalDateTime
 data class SearchCartQueryResponse(
     val id: String,
     val items: List<CartItemReadModel>,
-    val subTotal: Int,
-    val total: Int,
-    val discount: Int?,
+    val subTotal: Double,
+    val total: Double,
+    val discount: Double?,
     val updatedAt: LocalDateTime,
     val customerId: String?
 ) : QueryResponse
