+++ src/main/kotlin/com/sinaglife/shoping_cart/read_model/domain/CartReadModel.kt
@@ -5,9 +5,9 @@ import java.time.LocalDateTime
 data class CartReadModel(
     val id: String,
     val items: List<CartItemReadModel>,
-    val subTotal: Int,
-    val total: Int,
-    val discount: Int?,
+    val subTotal: Double,
+    val total: Double,
+    val discount: Double?,
     val createdAt: LocalDateTime,
     val updatedAt: LocalDateTime,
     val customerId: String?
